% Generated by build script
\name{ct_adverse_events}
\alias{ct_adverse_events}
\title{Fetch adverse events from a study's results}
\usage{
ct_adverse_events()
}
\arguments{
\item{nct_id}{NCT identifier}
}
\value{tibble: group_id, group_title, event_term, event_type,}
\description{Fetch adverse events from a study's results}
