% Generated by build script
\name{ct_field_values}
\alias{ct_field_values}
\title{Discover available values for a field}
\usage{
ct_field_values()
}
\arguments{
\item{field}{Field name (e.g. "OverallStatus", "Phase",}
}
\value{tibble: value, count}
\description{Discover available values for a field}
