% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.R
\name{ct_results}
\alias{ct_results}
\title{Fetch outcome measures from a study's results}
\usage{
ct_results(nct_id)
}
\arguments{
\item{nct_id}{NCT identifier}
}
\value{
tibble: group_id, group_title, measure_title, measure_type,
time_frame, value, spread, unit
}
\description{
Fetch outcome measures from a study's results
}
