% Generated by build script
\name{ct_results}
\alias{ct_results}
\title{Fetch outcome measures from a study's results}
\usage{
ct_results()
}
\arguments{
\item{nct_id}{NCT identifier}
}
\value{tibble: group_id, group_title, measure_title, measure_type,}
\description{Fetch outcome measures from a study's results}
